+++ CHANGE_LOG.md
@@ -22,6 +22,7 @@ Major changes:
  - Option.of() is now Option.some() and it no longer returns None if you pass it a None.
  - Added StaticImports.xformChars(CharSequence).
  - Added oneOf.OneOf2OrNone class.
+ - Added Xform.dropWhile()
 
 Here is a script to ease your upgrade from 2.1.1 to 3.0:
 ```bash